requirements.txt
openai-whisper==20231117
deep-translator==1.11.4
gTTS==2.5.1
torch==2.2.1
numpy==1.26.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.0
python-json-logger==2.0.7 